move out some of the extension files to specific control folder it actually needs it. some of the unused extension were deleted.
