+++ ios/FluentUI/Navigation/NavigationBar.swift
@@ -686,12 +686,25 @@ open class NavigationBar: UINavigationBar {
     // MARK: Large/Normal Title handling
 
     private func updateViewsForLargeTitlePresentation(for navigationItem: UINavigationItem?) {
+        // UIView.isHidden has a bug where a series of repeated calls with the same parameter can "glitch" the view into a permanent shown/hidden state
+        // i.e. repeatedly trying to hide a UIView that is already in the hidden state
+        // by adding a check to the isHidden property prior to setting, we avoid such problematic scenarios
         if showsLargeTitle {
-            backgroundView.safelyShow()
-            contentStackView.safelyShow()
+            if backgroundView.isHidden {
+                backgroundView.isHidden = false
+            }
+
+            if contentStackView.isHidden {
+                contentStackView.isHidden = false
+            }
         } else {
-            backgroundView.safelyHide()
-            contentStackView.safelyHide()
+            if !backgroundView.isHidden {
+                backgroundView.isHidden = true
+            }
+
+            if !contentStackView.isHidden {
+                contentStackView.isHidden = true
+            }
         }
         updateShadow(for: navigationItem)
     }
